+++ app/lib/l10n/gen/ouds_flutter_app_localizations.dart
@@ -869,6 +869,36 @@ abstract class AppLocalizations {
   /// **'Helper Text'**
   String get app_components_textInputHelperText_label;
 
+  /// No description provided for @app_components_phoneNumberInput_label.
+  ///
+  /// In en, this message translates to:
+  /// **'Phone number input'**
+  String get app_components_phoneNumberInput_label;
+
+  /// No description provided for @app_components_phoneNumberInput_description_text.
+  ///
+  /// In en, this message translates to:
+  /// **'A Phone Number Input is a form field specifically designed to capture and validate telephone numbers, often in international format. It typically integrates a country selector, allowing users to choose their country and automatically apply the corresponding dialing code (such as +33 for France).'**
+  String get app_components_phoneNumberInput_description_text;
+
+  /// No description provided for @app_components_phoneNumberInput_country_selector_label.
+  ///
+  /// In en, this message translates to:
+  /// **'Country selector'**
+  String get app_components_phoneNumberInput_country_selector_label;
+
+  /// No description provided for @app_components_phoneNumberInput_prefix_label.
+  ///
+  /// In en, this message translates to:
+  /// **'Prefix'**
+  String get app_components_phoneNumberInput_prefix_label;
+
+  /// No description provided for @app_components_phoneNumberInput_error_label.
+  ///
+  /// In en, this message translates to:
+  /// **'The phone number can\'t be empty.'**
+  String get app_components_phoneNumberInput_error_label;
+
   /// No description provided for @app_components_link_label.
   ///
   /// In en, this message translates to: