+++ app/lib/l10n/gen/ouds_flutter_app_localizations_en.dart
@@ -422,6 +422,24 @@ class AppLocalizationsEn extends AppLocalizations {
   String get app_components_textInputHelperText_label => 'Helper Text';
 
   @override
+  String get app_components_phoneNumberInput_label => 'Phone number input';
+
+  @override
+  String get app_components_phoneNumberInput_description_text =>
+      'A Phone Number Input is a form field specifically designed to capture and validate telephone numbers, often in international format. It typically integrates a country selector, allowing users to choose their country and automatically apply the corresponding dialing code (such as +33 for France).';
+
+  @override
+  String get app_components_phoneNumberInput_country_selector_label =>
+      'Country selector';
+
+  @override
+  String get app_components_phoneNumberInput_prefix_label => 'Prefix';
+
+  @override
+  String get app_components_phoneNumberInput_error_label =>
+      'The phone number can\'t be empty.';
+
+  @override
   String get app_components_link_label => 'Link';
 
   @override